requirements.txt
flask==3.0.2
flask-cors==4.0.0
langchain==0.3.19
langchain-core==0.3.37
langchain-openai==0.3.6
python-dotenv==1.0.1
openai==1.63.2
streamlit==1.32.2
requests==2.32.3
